API

Public

Dante.EulerExact — Method

EulerExact(ρ1, u1, p1, ρ4, u4, p4, tEnd, n)

Classical Gas Exact Riemann Solver for shock tube problems This programs is based on the code of P. Wesseling. PRINCIPLES OF COMPUTATIONAL FLUID DYNAMICS Springer-Verlag, Berlin etc., 2001. ISBN 3-540-67853-0 NOTE: A Cavitation Check is incorporated in the code. It further prevents plotting for possible but physically unlikely case of expansion shocks.

Arguments

ρ1::Float64: left density at t=0.

u1::Float64: left velocity at t=0.

p1::Float64: left pressure at t=0.

ρ4::Float64: right density at t=0.

u4::Float64: right velocity at t=0.

p4::Float64: right pressure at t=0.

tEnd::Float64: final solution time.

n::Integer64: the gas Degree of freedom.

Coded by Manuel Diaz, IAM, NTU 03/09/2011. Migrated by Hongyang Zhou from MATLAB to Julia, 11/05/2019

Dante.set_init_Riemann — Function

set_init_Riemann(RiemannProblemType, Verbose)

Set the initial conditions of Riemann problems. Note that currently tEnd and CFL number can only be set in PARAM.toml, and cannot be changed afterwards.

Dante.solve — Function

solve(paramFile)

Run the model given input parameter file paramFile.

Private

Dante.Param3D — Type

Model input parameters

Dante.advance! — Method

Explicit time advance.

Dante.calc_face_value! — Method

calc_face_value!(param, state_GV, faceState, faceGradient)

Type instability for the return type is introduced, but it seems ok. I don't know how to modify faceState for views in 1st order.

Dante.divergence! — Method

divergence!(param, vec, div)

Calculate the divergence of vectors specialized to my grid size. Always assume starting with i -> j -> k for 1/2/3D! Right now do nothing for the ghost cells. Maybe needed later! Take central differences on interior points.

Dante.divergence_ndgrid! — Method

divergence_ndgrid!(hx, hy, hz, vec, div)

Generic divergence div of vector vec with step lengths hx, hy, hz.

Dante.get_speed_max! — Method

Calculate the maximum speed in each direction.

Dante.minmod — Method

minmod(a, b, c)

For three inputs, use Harten's generalized definition. Return zero if opposite sign, otherwise the one of smaller magnitude.

Dante.minmod — Method

minmod(a, b)

Return zero if opposite sign, otherwise the one of smaller magnitude.

Dante.plot_Riemann_exact — Method

Plot the analytical solution of shock tube problem.

Dante.plotvar — Method

Plot 1D variables along a line.

Dante.setParameters — Method

setParameters(filename)

Read parameters from PARAM.toml and construct the parameter list.

Dante.update_state! — Method

Time-accurate state update.

Dante.update_state! — Method

Local timestepping state update.